Automatically backed up by Learn

+++ index.js
@@ -32,5 +32,5 @@ return parseInt(a, 10)
 }
 
 function makeInt(a) {
-  return parseInt('0x2328')
+  return parseFloat('0x2328')
 }
